MAJOR optimisation!
Now, each row is lazily loaded, including each row's elements.
Scrolling performance on big documents has been improved dramatically, by a clever idea: When a cell is not selected, display a simple Text composable, and when it is selected, convert it to a TextField. This saves a lot of resources as textfields are heavy to render. Every possible variable is cached, and the least amount of computation is ever done.